Refactor KanbanPlugin and ThreadPlugin to use `aurora-theme`’s `fragments`. Remove `ThemePlugin`’s `styles.ts` export, migrate those stories upstream to `aurora`.
