+++ packages/apps/plugins/plugin-kanban/src/components/KanbanMain.tsx
@@ -4,10 +4,9 @@
 
 import React, { FC } from 'react';
 
-import { Styles } from '@braneframe/plugin-theme';
 import { Kanban as KanbanType } from '@braneframe/types';
 import { Input, Main, useTranslation } from '@dxos/aurora';
-import { blockSeparator, mx } from '@dxos/aurora-theme';
+import { baseSurface, blockSeparator, mx } from '@dxos/aurora-theme';
 import { SpaceProxy } from '@dxos/client/echo';
 import { Text } from '@dxos/echo-schema';
 
@@ -32,7 +31,7 @@ export const KanbanMain: FC<{ data: { space: SpaceProxy; object: KanbanType } }>
   };
 
   return (
-    <Main.Content classNames={mx('flex flex-col grow min-bs-[100vh] overflow-hidden', Styles.level0.bg)}>
+    <Main.Content classNames={mx('flex flex-col grow min-bs-[100vh] overflow-hidden', baseSurface)}>
       <div>
         <Input.Root>
           <Input.Label srOnly>{t('kanban title label')}</Input.Label>
