+++ packages/apps/plugins/plugin-theme/src/index.ts
@@ -2,7 +2,6 @@
 // Copyright 2023 DXOS.org
 //
 
-export * from './styles';
 export * from './types';
 export * from './util';
