--- packages/apps/plugins/plugin-theme/src/styles.ts
@@ -1,25 +1,0 @@
-//
-// Copyright 2023 DXOS.org
-//
-
-/**
- * Cross-cutting styles used by all plugins.
- */
-// TODO(burdon): Dynamically provided by theme?
-export const Styles = {
-  level0: {
-    bg: 'bg-white dark:bg-neutral-900',
-    text: 'text-black dark:text-neutral-200',
-    border: 'border border-neutral-300 dark:border-neutral-800',
-  },
-  level1: {
-    bg: 'bg-neutral-50 dark:bg-neutral-800',
-    text: 'text-black dark:text-neutral-200',
-    border: 'border border-neutral-300 dark:border-neutral-800',
-  },
-  level2: {
-    bg: 'bg-neutral-100 dark:bg-neutral-700',
-    text: 'text-black dark:text-neutral-200',
-    border: 'border border-neutral-300 dark:border-neutral-800',
-  },
-};
